Move packages (root and eventselector into api).
