+++ src/main/scala/freeeventsourcing/akka/PersistenceIds.scala
@@ -1,7 +1,7 @@
 package freeeventsourcing.akka
 
 import freeeventsourcing.utils.{ CompositeName, StringSerializable }
-import freeeventsourcing.{ Aggregate, BoundedContext }
+import freeeventsourcing.api.{ Aggregate, BoundedContext }
 
 object PersistenceIds {
   def forAggregate[A <: Aggregate](boundedContext: BoundedContext, aggregate: A)(id: A#Id)(implicit ser: StringSerializable[A#Id]): String = {
