+++ src/main/scala/freeeventsourcing/api/AggregateImplementation.scala
@@ -1,4 +1,4 @@
-package freeeventsourcing
+package freeeventsourcing.api
 
 import scala.language.implicitConversions
 import scala.collection.immutable.Seq
