+++ src/main/scala/freeeventsourcing/api/EventSelector.scala
@@ -1,10 +1,10 @@
-package freeeventsourcing
+package freeeventsourcing.api
 
 import scala.language.implicitConversions
 import scala.collection.immutable.Set
-import freeeventsourcing.EventSelector.ops._
 import freeeventsourcing.utils.CompositeName
 import simulacrum.typeclass
+import EventSelector.ops._
 
 /** Selects a specific subset of events. */
 @typeclass trait EventSelector[S <: EventSelector.WithEventType] {
