+++ src/main/scala/freeeventsourcing/api/package.scala
@@ -3,7 +3,7 @@ package freeeventsourcing
 import scala.annotation.implicitNotFound
 import shapeless.ops.hlist.Selector
 
-package object support {
+package object api {
   @implicitNotFound("Aggregate ${A} is not a member of this bounded context.")
   type ValidAggregate[BC <: BoundedContext, A <: Aggregate] = Selector[BC#Aggregates, A]
 }
