+++ src/main/scala/freeeventsourcing/syntax/MonadicCommandHandlerSyntax.scala
@@ -6,7 +6,7 @@ import scala.collection.immutable.Seq
 import cats.Monad
 import cats.data.{ WriterT, Xor }
 import cats.instances.list._
-import freeeventsourcing.AggregateCommand
+import freeeventsourcing.api.AggregateCommand
 import shapeless.ops.coproduct.Inject
 
 /** Syntactic support for writing command handlers in monadic style. */
