+++ src/test/scala/freeeventsourcing/accountprocessing/AccountProcessing.scala
@@ -2,7 +2,7 @@ package freeeventsourcing.accountprocessing
 
 import shapeless.{ ::, HNil }
 import freeeventsourcing.accountprocessing.Account.Event.Closed
-import freeeventsourcing.BoundedContext
+import freeeventsourcing.api.BoundedContext
 
 object AccountProcessing extends BoundedContext {
   val name = "account processing"
