+++ src/test/scala/freeeventsourcing/api/AggregateEventTests.scala
@@ -1,8 +1,8 @@
-package freeeventsourcing
+package freeeventsourcing.api
 
-import freeeventsourcing.ProcessActionTests.{ OtherEvent, TestAggregate }
 import freeeventsourcing.accountprocessing.Account.Event.Opened
 import freeeventsourcing.accountprocessing._
+import freeeventsourcing.api.ProcessActionTests.{ OtherEvent, TestAggregate }
 import org.scalatest.{ FlatSpec, Matchers }
 
 class AggregateEventTests extends FlatSpec with Matchers {
@@ -13,6 +13,6 @@ class AggregateEventTests extends FlatSpec with Matchers {
 
   "AggregateEvent " should " have eventType equal to the absolute classname of the event when not nested in the aggregate" in {
     val e = AggregateEvent[TestAggregate.type, OtherEvent](TestAggregate, TestAggregate.Id(1), OtherEvent())
-    e.eventType shouldBe "freeeventsourcing.ProcessActionTests$OtherEvent"
+    e.eventType shouldBe "freeeventsourcing.api.ProcessActionTests$OtherEvent"
   }
 }
