+++ src/test/scala/freeeventsourcing/api/ProcessDefinitionTests.scala
@@ -1,10 +1,10 @@
-package freeeventsourcing
+package freeeventsourcing.api
 
 import freeeventsourcing.accountprocessing.Account.Command.BlockFunds
 import freeeventsourcing.accountprocessing.Account.Error.{ InsufficientFunds, NotOpen }
 import freeeventsourcing.accountprocessing.Transaction.Event.Created
-import freeeventsourcing.accountprocessing.{ Account, AccountProcessing, Transaction }
-import freeeventsourcing.eventselector.{ AggregateEventSelector, AggregateTypeEventSelector }
+import freeeventsourcing.accountprocessing.{ AccountProcessing, Transaction }
+import freeeventsourcing.api.eventselector.AggregateTypeEventSelector
 import freeeventsourcing.syntax.ProcessSyntax
 import org.scalatest.{ FlatSpec, Matchers }
