+++ src/test/scala/freeeventsourcing/api/eventselector/AggregateEventSelectorTests.scala
@@ -1,13 +1,14 @@
-package freeeventsourcing.eventselector
+package freeeventsourcing.api.eventselector
 
-import freeeventsourcing.EventSelector.WithEventType
-import freeeventsourcing.EventSelector.ops._
-import freeeventsourcing.EventTime.Zero
+import freeeventsourcing.api.EventSelector.WithEventType
+import freeeventsourcing.api.EventSelector.ops._
+import freeeventsourcing.api.EventTime.Zero
 import freeeventsourcing.accountprocessing.Account.Event.{ Closed, Opened }
 import freeeventsourcing.accountprocessing.Transaction.Event.Created
 import freeeventsourcing.accountprocessing._
 import freeeventsourcing.utils.CompositeName
 import freeeventsourcing._
+import freeeventsourcing.api._
 import org.scalatest.{ FlatSpec, Matchers }
 import shapeless.{ ::, HNil }
