+++ src/test/scala/freeeventsourcing/syntax/CoproductCommandHandlerTests.scala
@@ -1,7 +1,7 @@
 package freeeventsourcing.syntax
 
 import cats.data.Xor
-import freeeventsourcing.AggregateCommand
+import freeeventsourcing.api.AggregateCommand
 import org.scalatest.{ FlatSpec, Matchers }
 import shapeless.{ :+:, CNil, Coproduct }
